Download latest phpunit results cache before parallel tests
The PHPUnit parallel test implementation has support for balancing
buckets by duration when timing information is provided for the
test classes.

Update the `phpunit:prepare-parallel:...` action to download the
latest phpunit results cache file and process it as part of the
test suite bucket / `split_group` creation process. If the file
is empty or cannot be download, proceed as before - simply put
the same number of test classes in each bucket.

Bug: T378797
Bug: T384927

+++ includes/composer/PhpUnitSplitter/TestDescriptor.php
@@ -12,13 +12,13 @@ class TestDescriptor {
 	private string $className;
 	private array $namespace;
 	private ?string $filename;
-	private int $duration;
+	private float $duration;
 
 	public function __construct(
 		string $className,
 		array $namespace,
 		?string $filename = null,
-		int $duration = 0
+		float $duration = 0
 	) {
 		$this->className = $className;
 		$this->namespace = $namespace;
@@ -46,7 +46,7 @@ class TestDescriptor {
 		return implode( '\\', $this->namespace ) . '\\' . $this->className;
 	}
 
-	public function getDuration(): int {
+	public function getDuration(): float {
 		return $this->duration;
 	}
